refactor(editor): split RichTextEditor into focused modules
RichTextEditor.tsx had grown into a 465-line single file that mixed
TipTap setup, a full formatting toolbar with inline callout dropdown,
three media URL prompts, and image upload wiring for the button, drop
handler, and paste handler.

Split the component into focused modules:

- useImageUpload — shared upload state + drop/paste/toolbar handlers so
  the uploading spinner and error toast stay consistent regardless of
  how the user supplies the image.
- useMediaPrompts — link/image/youtube/audio prompt flows, isolated
  from the render layer so each can be adjusted without touching the
  editor wiring.
- CalloutDropdown — self-contained dropdown that owns its own
  open/close + click-outside handling.
- EditorToolbar — the entire formatting toolbar, parameterised on the
  editor and a small handler interface.
- RichTextEditor — thin composition that builds the TipTap editor and
  stitches the pieces together.

No behavior changes: same toolbar layout, same prompts, same
extensions, same image upload flows. All 85 frontend tests pass, lint
clean, type-check clean, production build ok.

Made-with: Cursor
